analysis/requirements.txt
requests
python-dotenv
pyyaml
